service_ai/requirement.txt
torch>=2.0.0
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
Pillow>=9.3.0
pyyaml>=6.0
seaborn>=0.12.0
opencv-python>=4.10.0.84
tqdm>=4.65.0
scipy>=1.10.0
requests>=2.28.0
ultralytics>=8.0.0
flask>=3.0.0
werkzeug>=3.0.0
paho-mqtt